+++ test/fixtures/resolveDirectory/entry.js
@@ -1,0 +1,5 @@
+import Test from "./resDir.svelte";
+
+new Test({
+    target: document.body,
+});