+++ test/fixtures/resolveDirectory/resDir.svelte
@@ -1,0 +1,14 @@
+<script>
+    // your script goes here
+    let x = 5;
+</script>
+
+<style lang="scss">
+    div {
+        background-image: url('fakeImage.png');
+    }
+</style>
+
+<div>
+    <h1>Hello World</h1>
+</div>